+++ tests/jni_api.rs
@@ -580,6 +580,18 @@ pub fn new_direct_byte_buffer() {
 }
 
 #[test]
+pub fn new_direct_byte_buffer_raw() {
+    let env = attach_current_thread();
+    let mut vec: Vec<u8> = vec![0, 1, 2, 3];
+    let buf = vec.as_mut_slice();
+    let result = unsafe {
+        env.new_direct_byte_buffer_raw(vec.as_mut_ptr(), vec.len())
+    };
+    assert!(result.is_ok());
+    assert!(!result.unwrap().is_null());
+}
+
+#[test]
 pub fn get_direct_buffer_capacity_ok() {
     let env = attach_current_thread();
     let mut vec: Vec<u8> = vec![0, 1, 2, 3];